adding final structural changes
