+++ general/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class GeneralConfig(AppConfig):
+    name = 'general'
