+++ general/models.py
@@ -1,0 +1,3 @@
+from django.db import models
+
+# Create your models here.
