+++ general/templates/start_screen.html
@@ -44,7 +44,7 @@
                     "source": "start_screen"
                 }   
                 setChannels(channel_options);
-                create_invite_link('.copy-invite'); 
+                create_invite_link('.copy-invite', party_type); 
 
                 // If the party is full and all players have joined start
                 // timer to redirect to game
