+++ trivia_party/urls.py
@@ -3,19 +3,8 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-     path('', views.new_party, name='new_party'),
-     path('join/find_player', views.find_player, name='find_player'),
-     path('join/create_player', views.create_player, name='create_player'),
-     path('join/<str:party_id>', views.join_party, name='join_party'),
-     path('join/<str:party_id>', views.join_party, name='join_party'),
-     path('start/<str:party_id>', views.start_screen, name='start_screen'),
-     path('finish/<str:party_id>', views.finish_screen, name='finish_screen'),
-     path('waiting/<str:party_id>', views.waiting_screen, name='waiting_screen'),
-     path('create_or_join_party', views.create_or_join_party,
-          name='create_or_join_party'),
-     path('recreate_party', views.recreate_party,
-          name='recreate_party'),
-     path('trivia/<str:party_id>', views.trivia_party, name='trivia_party'),
-     path('trivia/<str:party_id>/submission', 
-               views.submit_question, name='submission'),
+     path('<str:party_id>/finish', views.finish_screen, name='finish_screen'),
+     path('<str:party_id>/waiting', views.waiting_screen, name='waiting_screen'),
+     path('<str:party_id>/party', views.trivia_party, name='trivia_party'),
+     path('<str:party_id>/submission', views.submit_question, name='submission'),
 ]
